Minor changes to the unis panel
Now the list of universities has a nice heading and is contained in a
panel. I changed the cursor to pointer to show some feedback to the
user.

I also tried to change the pointers of the nodes in the svg to crosshair
(as you can probably see), to no avail. Good luck with it!

I encourage you from the very bottom of my heart to stop writing your
css and javascript in html. And, for God's sake, stop writing raw html
tags as javascript strings!!!!!!!!!

:beers: